backend/requirements.txt
Flask>=2.0
pyodbc>=4.0
python-dotenv>=1.0